--- app/client/App.js
@@ -1,16 +1,0 @@
-import React from 'react'
-
-export default class App extends React.Component {
-  constructor(props) {
-    super(props)
-  }
-
-  render() {
-    return (
-      <div>
-        {this.props.children}
-      </div>
-    )
-  }
-
-}
