+++ app/client/pages/configure-accounts.jsx
@@ -1,5 +1,6 @@
 import React from 'react'
 
 export default class ConfigureAccounts extends React.Component{
-  render = () => <div>config</div>
+  render = () =>
+    <div>config</div>
 }
